Reduce complexity of `generate_container_count`
Rather than calling `highlight_str` a bunch, use `format!` to assemble
the string. This also removes the need to `use std::convert::TryInto`,
since I'm `as isize`ing everywhere. It also respects and updates
`available_space`, though I'm not quite sure if a string as short as
this should do so.