Pass FlatJson and Row into LinePrinter; have it figure out Label stuff
With the introduction of YAML support in 2401816eacf4d4ba, we now have to
support non-string keys for Objects. Non-string keys should be surrounded with
square brackets, rather than quotes, and we should always show these square
brackets, even in Data mode.

This branch started trying to remedy that, but I ran into an issue: the
ScreenWriter was responsible for passing the label into the LinePrinter, and it
would handle stripping out the delimiter, whether that was quotes or square
brackets. And, unless the value of the line was a container, the LinePrinter
wouldn't have access to the FlatJson, and the full pretty-printed
representation, to determine on its own what the correct delimiter was. Also I
was having trouble keeping the indexes straight when highlighting stuff with
and without delimiters -- lots of off-by-one errors.

I could shove additional logic into the ScreenWriter, but this felt weird. We
already sometimes pass in the FlatJson to the LinePrinter for generating
container previews, so why not just pass it in all the time? Originally I had
hoped for the LinePrinter to only need minimal context, but that ship has
clearly sailed.


Rather than try to fix the non-string keys right away, I've decided to move
more stuff into LinePrinter, and I've added some YAML tests that exhibit the
_incorrect_ behavior. Hopefully after this it'll be easier to fix the
delimiter issue.

I incrementally made the following changes here:
- Always pass FlatJson into LinePrinter as one of its fields, and remove it
  from the LineLabel::Container. A bunch of functions passed in a FlatJson that
  could now just use the one on `self`.
- Also pass in the Row the LinePrinter is printing out. This change mostly
  makes sense! Now that we have the Row, we can remove the
  used-in-exactly-one-place `node_depth`.
- I thought passing in both `depth` and `tab_size` was stupid, so I have the
  ScreenWriter (which knows about indentation reductions via `<`) calculate the
  correct indentation and pass that in.
- Then I removed the `label` and `label_range` fields from LinePrinter, and
  just had it figure those out on its own inside `fill_in_label`. This moved
  code is what will be modified to fix the delimiter issue.
- Moved some label highlighting logic to its own function as well.
- Updated all the tests.

In the end, this has just shoved more logic into LinePrinter -- already a very
complicated file -- so not really sure how much of a win there is here.