use a channel for stdin and SIGWINCH
By using a channel, we can remove the need for calling libc::poll on
the stdin pipe. This doesn't work on the Mac: when we redirect tty to
stdin, it breaks the original stdin pipe (fd 0), such that libc::poll on
it returns instantly, with an revents of POLLNVAL (invalid pipe). macOS
also doesn't support libc::poll on devices, so we can't poll tty
directly.

Instead, we spawn two threads: one just reads stdin forever, and the
other uses the socket to listen for SIGWINCH events as before (using
libc::poll). These both write to a channel, and the TuiInput simply
reads from this channel.

This fixes #2.